chatbot-backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
python-dotenv==1.1.0
pymongo==4.11.3
langchain==0.3.21
langchain-community==0.3.20
langchain-pinecone==0.2.3
openai==1.68.2
pinecone==5.4.0
sentence-transformers==3.4.1
gunicorn==23.0.0
eventlet==0.39.1
flask-cors==5.0.1
